# src/mdm.api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "行動裝置管理系統 API — 裝置管控、資產管理、借用流程、App 派發",
        "title": "MDM Server API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "分類列表（樹狀排序）/ 新增分類",
                "parameters": [
                    {
                        "description": "新增分類（POST）",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.swagCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET: {categories: [...]}, POST: {id: ...}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/controller.swagCategoryItem"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "分類列表（樹狀排序）/ 新增分類",
                "parameters": [
                    {
                        "description": "新增分類（POST）",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.swagCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET: {categories: [...]}, POST: {id: ...}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/controller.swagCategoryItem"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "更新 / 刪除分類",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分類 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新名稱（PUT）",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.swagCategoryUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagOK"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "更新 / 刪除分類",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分類 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新名稱（PUT）",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.swagCategoryUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagOK"
                        }
                    }
                }
            }
        },
...
        "/api/devices-available": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "可借用裝置列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagDeviceAvailResp"
                        }
                    }
                }
            }
        },
        "/api/devices-list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "裝置列表（含資產與借用狀態）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "關鍵字篩選",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分類 ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "保管人 ID",
                        "name": "custodian_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "借用狀態",
                        "name": "rental_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagDeviceListResp"
                        }
                    }
                }
            }
        },
        "/api/devices/{udid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "依 UDID 取得裝置詳細資訊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "裝置 UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagDevice"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            }
        },
...
        "/api/sync-device-info": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "向所有裝置發送 DeviceInformation 查詢",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagSyncCountResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.swagCategoryItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
        "controller.swagCategoryReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                }
            }
        },
        "controller.swagCategoryUpdateReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "controller.swagDevice": {
            "type": "object",
            "properties": {
                "battery_level": {
                    "type": "number"
                },
                "device_name": {
                    "type": "string"
                },
                "enrollment_status": {
                    "type": "string"
                },
                "is_lost_mode": {
                    "type": "boolean"
                },
                "is_supervised": {
                    "type": "boolean"
                },
                "last_seen": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "os_version": {
                    "type": "string"
                },
                "serial_number": {
                    "type": "string"
                },
                "udid": {
                    "type": "string"
                }
            }
        },
        "controller.swagDeviceAvailItem": {
            "type": "object",
            "properties": {
                "asset_status": {
                    "type": "string"
                },
                "device_name": {
                    "type": "string"
                },
                "enrollment_status": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "os_version": {
                    "type": "string"
                },
                "serial_number": {
                    "type": "string"
                },
                "udid": {
                    "type": "string"
                }
            }
        },
        "controller.swagDeviceAvailResp": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.swagDeviceAvailItem"
                    }
                }
            }
        },
        "controller.swagDeviceListItem": {
            "type": "object",
            "properties": {
                "asset_status": {
                    "type": "string"
                },
                "battery_level": {
                    "type": "number"
                },
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "custodian_id": {
                    "type": "string"
                },
                "custodian_name": {
                    "type": "string"
                },
                "device_name": {
                    "type": "string"
                },
                "enrollment_status": {
                    "type": "string"
                },
                "is_lost_mode": {
                    "type": "boolean"
                },
                "is_supervised": {
                    "type": "boolean"
                },
                "last_seen": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "os_version": {
                    "type": "string"
                },
                "serial_number": {
                    "type": "string"
                },
                "udid": {
                    "type": "string"
                }
            }
        },
        "controller.swagDeviceListResp": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.swagDeviceListItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "controller.swagError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "controller.swagOK": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "controller.swagSyncCountResp": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}